Consolidate get_sock_opt and set_sock_opt tests, remove google.com dependency
